feat(provider): mvvm pattern
