+++ food_recipe/lib/presentation/saved_recipes/view/view.dart
@@ -1,0 +1,1 @@
+export 'saved_recipes.dart';
