+++ food_recipe/lib/presentation/sign_up/view/view.dart
@@ -1,0 +1,1 @@
+export 'sign_up.dart';
